Split SCMStepTest into 3 tests
Isolate Git, Subversion, and others
